refactor: Changed how linting is called
For the Molecule scripts I want them all sharing the same thing so I
created a script they call.  That way any changes of linting tooling
only needs to be modified in that script.

Also added helper files for linting and testing that will use poetry and
the virtual environment to run the scripts.